[refactor] Excel Helper & Separate Logic
