+++ ContactsTracker/Data/DataEntry.cs
@@ -73,7 +73,7 @@ public class DataEntry(string? territoryName, string? rouletteType, bool isCompl
                     if (partyMember != null)
                     {
                         var worldID = groupManager->GetPartyMemberByContentId((ulong)partyMember.ContentId)->HomeWorld;
-                        var worldName = Plugin.DataManager.GetExcelSheet<World>()?.GetRow(worldID).Name.ExtractText();
+                        var worldName = ExcelHelper.GetWorldName(worldID);
                         names[i] = $"{partyMember.Name} @ {worldName}";
                         if (configuration.LogPartyClass)
                         {
